+++ pg_documentdb/src/test/regress/sql/public_api_schema.sql
@@ -1,4 +1,4 @@
--- show all functions exported in documentdb_api.
+-- show all functions, procedures, aggregates and window aggregates exported in documentdb_api.
 \df documentdb_api.*
 
 \df documentdb_api_catalog.*
@@ -6,15 +6,6 @@
 \df documentdb_api_internal.*
 
 \df documentdb_data.*
-
--- show all aggregates exported
-\da+ documentdb_api.*
-
-\da+ documentdb_api_catalog.*
-
-\da+ documentdb_api_internal.*
-
-\da+ documentdb_data.*
 
 -- Access methods + Operator families
 \dA *documentdb*