+++ scripts/validate_extension_file_state.sh
@@ -20,7 +20,7 @@ function ValidateNoCitusReferences()
 
     # TODO: Clean these up
     validationExceptions="documentdb--0.24-0.sql create_index_background--0.23-0.sql create_index_background--latest.sql drop_indexes.c create_indexes_background.c"
-    validationExceptions="$validationExceptions current_op.c index.c ttl_support_functions--0.24-0.sql ttl_support_functions--latest.sql"
+    validationExceptions="$validationExceptions current_op.c index.c ttl_support_functions--0.24-0.sql"
     validationExceptions="$validationExceptions diagnostic_commands_common.c"
 
     foundInvalid="";